优化获取应用服务方法
