+++ common.props
@@ -7,7 +7,7 @@
 
     <LicenseUrl>https://raw.githubusercontent.com/liuhll/silky/main/LICENSE</LicenseUrl>
 
-    <Version>3.0.5-bate8</Version>
+    <Version>3.0.5-bate11</Version>
 
     <PackageIconUrl>https://raw.githubusercontent.com/liuhll/silky/main/docs/.vuepress/public/assets/logo/logo.png</PackageIconUrl>
