+++ framework/src/Silky.Core/Configuration/AppSettingsOptions.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace Silky.Core.Configuration
 {
     public class AppSettingsOptions
@@ -8,8 +10,7 @@ namespace Silky.Core.Configuration
         {
             DisplayFullErrorStack = false;
             AutoValidationParameters = true;
-            AppServiceInterfacePattern = "Application.Contracts";
-            AppServicePattern = "Application";
+            Services = new List<AppServiceOptions>();
         }
 
         /// <summary>
@@ -28,24 +29,6 @@ namespace Silky.Core.Configuration
         /// </summary>
         public bool AutoValidationParameters { get; set; }
 
-        /// <summary>
-        /// 应用服务接口扫描的目录
-        /// </summary>
-        public string AppServiceInterfaceDirectory { get; set; }
-
-        /// <summary>
-        /// 应用服务接口正则匹配模式
-        /// </summary>
-        public string AppServiceInterfacePattern { get; set; }
-
-        /// <summary>
-        /// 应用服务接口扫描的目录
-        /// </summary>
-        public string AppServiceDirectory { get; set; }
-
-        /// <summary>
-        /// 应用服务正则匹配模式
-        /// </summary>
-        public string AppServicePattern { get; set; }
+        public ICollection<AppServiceOptions> Services { get; set; }
     }
 }
