+++ framework/test/GatewayDemo/appsettings.yaml
@@ -19,8 +19,9 @@ rpc:
 #  instanceEnabled: true
 #  namingUseRpc: true
 appSettings:
-  appServiceInterfaceDirectory: ..
-  appServiceInterfacePattern: I*Application
+  services:
+    - appServiceInterfaceDirectory: ..
+      appServiceInterfacePattern: I*Application
 
 registrycenter:
   type: Consul
